Move test under feature package
