+++ src/test/java/test/skip/github1632/SkipMonitoringListener.java
@@ -1,4 +1,4 @@
-package test.listeners.issue1632;
+package test.skip.github1632;
 
 import java.util.HashMap;
 import java.util.Map;
