+++ src/test/java/test/skip/github1632/TestClassSample.java
@@ -1,4 +1,4 @@
-package test.listeners.issue1632;
+package test.skip.github1632;
 
 import org.testng.Assert;
 import org.testng.annotations.Test;
