+++ src/test/resources/testng.xml
@@ -186,7 +186,7 @@
       <class name="test.listeners.github1735.ExecutionListenerTest"/>
       <class name="test.listeners.github1602.IssueTest"/>
       <class name="test.listeners.issue1777.IssueTest"/>
-      <class name="test.listeners.issue1632.IssueTest"/>
+      <class name="test.skip.github1632.IssueTest"/>
       <class name="test.listeners.issue2220.IssueTest"/>
       <class name="test.listeners.issue2328.IssueTest"/>
       <class name="test_result.GitHub1197Test"/>
